// bank-ui/src/app/models/user.ts
export class User {
    id?: string;
    username?: string;
    password?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    customerNumber?: string;
    accountNumber?: string;
    accessToken?: string;
}

// export class UserInfo {
//     "id": 2,
//     "firstName": null,
//     "lastName": null,
//     "customerNumber": 2966880461,
//     "status": "Active",
//     "address": null,
//     "email": "ngoswami@deqode.com",
//     "createDateTime": "11:30:48",
//     "updateDateTime": null,
//     "account": {
//         "id": 1,
//         "accountNumber": 8580318151,
//         "accountStatus": "Active",
//         "accountType": "Saving",
//         "accountBalance": 0,
//         "createDateTime": "11:30:48",
//         "updateDateTime": "11:30:48"
//     }
// }